# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Academix REST API
  version: 1.0.0
  description: >-
    Contrato final del backend. React consume únicamente esta API; las claves
    administrativas y las tablas de Supabase nunca se exponen al navegador.
servers:
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
...
security: []
paths:
...
  /courses/{courseId}/reviews:
    parameters: [{ $ref: "#/components/parameters/CourseId" }]
    get:
      tags: [Reviews]
      operationId: listCourseReviews
      responses:
        "200": { $ref: "#/components/responses/Paginated" }
    post:
      tags: [Reviews]
      operationId: createCourseReview
      security: [{ bearerAuth: [] }]
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
  /admin/users/{userId}/roles:
    parameters: [{ $ref: "#/components/parameters/UserId" }]
    patch:
      tags: [Administration]
      operationId: setUserRoles
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
...
  /admin/instructors/{userId}:
    parameters: [{ $ref: "#/components/parameters/UserId" }]
    put:
      tags: [Administration]
      operationId: upsertInstructor
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /admin/categories:
    post:
      tags: [Administration]
      operationId: createCategory
      security: [{ bearerAuth: [] }]
      responses:
        "201": { $ref: "#/components/responses/Success" }
  /admin/categories/{categoryId}:
    parameters: [{ $ref: "#/components/parameters/CategoryId" }]
    patch:
      tags: [Administration]
      operationId: updateCategory
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
...
  /admin/courses:
    get:
      tags: [Administration]
      operationId: adminListCourses
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Paginated" }
    post:
      tags: [Administration]
      operationId: createCourse
      security: [{ bearerAuth: [] }]
      responses:
        "201": { $ref: "#/components/responses/Success" }
...
  /admin/courses/{courseId}/instructor:
    parameters: [{ $ref: "#/components/parameters/CourseId" }]
    put:
      tags: [Administration]
      operationId: assignPrincipalInstructor
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /admin/courses/{courseId}/publish:
    parameters: [{ $ref: "#/components/parameters/CourseId" }]
    post:
      tags: [Administration]
      operationId: publishCourse
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
  /admin/courses/{courseId}/archive:
    parameters: [{ $ref: "#/components/parameters/CourseId" }]
    post:
      tags: [Administration]
      operationId: archiveCourse
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
...
  /admin/reviews/{reviewId}/moderation:
    parameters: [{ $ref: "#/components/parameters/ReviewId" }]
    patch:
      tags: [Administration]
      operationId: moderateReview
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourseId: { name: courseId, in: path, required: true, schema: { type: string, format: uuid } }
...
    UserId: { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
    CategoryId: { name: categoryId, in: path, required: true, schema: { type: string, format: uuid } }
...
    ReviewId: { name: reviewId, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { const: true }
        data: {}
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: true }
  responses:
    Success:
      description: Operación correcta
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessEnvelope" }
    Paginated:
      description: Colección paginada
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessEnvelope" }
    Error:
      description: Error controlado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...